Exam pepar generator/requirements.txt
flask==3.0.2
google-generativeai==0.3.2
python-dotenv==1.0.1
flask-sqlalchemy==3.1.1
flask-login==0.6.3
werkzeug==3.0.1 